Cells

RecurrentLayers.RANCell — Type

RANCell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

The RANCell, introduced in this paper, is a recurrent cell layer which provides additional memory through the use of gates.

and returns both ht anf ct.

See RAN for a layer that processes entire sequences.

Arguments

in => out: Specifies the input and output dimensions of the layer.

init: Initialization function for the weight matrices, default is glorot_uniform.

bias: Indicates if a bias term is included; the default is true.

Forward

rancell(x, [h, c])

The forward pass takes the following arguments:

x: Input to the cell, which can be a vector of size in or a matrix of size in x batch_size.

h: The hidden state vector of the cell, sized out, or a matrix of size out x batch_size.

c: The candidate state, sized out, or a matrix of size out x batch_size.

If not provided, both h and c default to vectors of zeros.

Examples

rancell = RANCell(3 => 5)
inp = rand(Float32, 3)
#initializing the hidden states, if we want to provide them
state = rand(Float32, 5)
c_state = rand(Float32, 5)

#result with default initialization of internal states
result = rancell(inp)
#result with internal states provided
result_state = rancell(inp, (state, c_state))

RecurrentLayers.IndRNNCell — Type

function IndRNNCell((in, out)::Pair, σ=relu;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.LightRUCell — Type

LightRUCell((in, out)::Pair, σ=tanh;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.LiGRUCell — Type

LiGRUCell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.MGUCell — Type

MGUCell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.NASCell — Type

NASCell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.RHNCell — Type

RHNCell((in, out), depth=3; couple_carry::Bool = true, cell_kwargs...)

RecurrentLayers.RHNCellUnit — Type

RHNCellUnit((in, out)::Pair;
    kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.MUT1Cell — Type

MUT1Cell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.MUT2Cell — Type

MUT2Cell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.MUT3Cell — Type

MUT3Cell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true)

RecurrentLayers.SCRNCell — Type

SCRNCell((in, out)::Pair;
    kernel_init = glorot_uniform,
    recurrent_kernel_init = glorot_uniform,
    bias = true,
    alpha = 0.0)