MinimalRNNCell

RecurrentLayers.MinimalRNNCell — Type

MinimalRNNCell(input_size => hidden_size;
    init_encoder_kernel = glorot_uniform,
    init_recurrent_kernel = glorot_uniform,
    init_memory_kernel = glorot_uniform,
    encoder_bias = true, recurrent_bias = true, memory_bias=true,
    independent_recurrence = false, integration_mode = :addition)

Minimal recurrent neural network unit (Chen, 2017). See MinimalRNN for a layer that processes entire sequences.

Arguments

input_size => hidden_size: input and inner dimension of the layer.

Keyword arguments

init_encoder_kernel: initializer for the input to hidden weights. Default is glorot_uniform.

init_recurrent_kernel: initializer for the hidden to hidden weights. Default is glorot_uniform.

init_memory_kernel: initializer for the memory to hidden weights. Default is glorot_uniform.

encoder_bias: include a bias in the encoder or not. Default is true.

recurrent_bias: include recurrent to recurrent bias or not. Default is true.

memory_bias: include memory to recurrent bias or not. Default is true.

independent_recurrence: flag to toggle independent recurrence. If true, the recurrent to recurrent weights are a vector instead of a matrix. Default false.

integration_mode: determines how the input and hidden projections are combined. The options are :addition and :multiplicative_integration. Defaults to :addition.

Equations

\[\begin{aligned} \mathbf{z}(t) &= \Phi(\mathbf{x}(t)) = \tanh\left( \mathbf{W}_{xz} \mathbf{x}(t) + \mathbf{b}^{z} \right), \\ \mathbf{u}(t) &= \sigma\left( \mathbf{W}_{hh}^{u} \mathbf{h}(t-1) + \mathbf{W}_{zh}^{u} \mathbf{z}(t) + \mathbf{b}^{u} \right), \\ \mathbf{h}(t) &= \mathbf{u}(t) \circ \mathbf{h}(t-1) + \left(1 - \mathbf{u}(t)\right) \circ \mathbf{z}(t) \end{aligned}\]

Forward

minimalrnncell(inp, state)
minimalrnncell(inp)

Arguments

inp: The input to the minimalrnncell. It should be a vector of size input_size or a matrix of size input_size x batch_size.

(state, cstate): A tuple containing the hidden and cell states of the MinimalRNNCell. They should be vectors of size hidden_size or matrices of size hidden_size x batch_size. If not provided, they are assumed to be vectors of zeros, initialized by Flux.initialstates.

Returns

A tuple (output, state), where output = new_state is the new hidden state and state = (new_state, new_cstate) is the new hidden and cell state. They are tensors of size hidden_size or hidden_size x batch_size.