MUT3Cell

RecurrentLayers.MUT3Cell — Type

MUT3Cell(input_size => hidden_size;
    init_kernel = glorot_uniform,
    init_recurrent_kernel = glorot_uniform,
    bias = true)

Mutated unit 3 cell (Jozefowicz et al., 07–09 Jul 2015). See MUT3 for a layer that processes entire sequences.

Arguments

input_size => hidden_size: input and inner dimension of the layer.

Keyword arguments

init_kernel: initializer for the input to hidden weights. Default is glorot_uniform.

init_recurrent_kernel: initializer for the hidden to hidden weights. Default is glorot_uniform.

bias: include input to recurrent bias or not. Default is true.

recurrent_bias: include recurrent to recurrent bias or not. Default is true.

independent_recurrence: flag to toggle independent recurrence. If true, the recurrent to recurrent weights are a vector instead of a matrix. Default false.

integration_mode: determines how the input and hidden projections are combined. The options are :addition and :multiplicative_integration. Defaults to :addition.

Equations

\[\begin{aligned} \mathbf{z}(t) &= \sigma\left( \mathbf{W}^{z}_{ih} \mathbf{x}(t) + \mathbf{W}^{z}_{hh} \tanh\left( \mathbf{h}(t) \right) + \mathbf{b}^{z} \right), \\ \mathbf{r}(t) &= \sigma\left( \mathbf{W}^{r}_{ih} \mathbf{x}(t) + \mathbf{W}^{r}_{hh} \mathbf{h}(t) + \mathbf{b}^{r} \right), \\ \mathbf{h}(t+1) &= \left[ \tanh\left( \mathbf{W}^{h}_{hh} \left( \mathbf{r}(t) \odot \mathbf{h}(t) \right) + \mathbf{W}^{h}_{ih} \mathbf{x}(t) + \mathbf{b}^{h} \right) \right] \odot \mathbf{z}(t) \\ &\quad + \mathbf{h}(t) \odot \left( 1 - \mathbf{z}(t) \right) \end{aligned}\]

Forward

mutcell(inp, state)
mutcell(inp)

Arguments

inp: The input to the mutcell. It should be a vector of size input_size or a matrix of size input_size x batch_size.

state: The hidden state of the MUTCell. It should be a vector of size hidden_size or a matrix of size hidden_size x batch_size. If not provided, it is assumed to be a vector of zeros, initialized by Flux.initialstates.

Returns

A tuple (output, state), where both elements are given by the updated state new_state, a tensor of size hidden_size or hidden_size x batch_size.