RANCell

RecurrentLayers.RANCell — Type

RANCell(input_size => hidden_size;
    init_kernel = glorot_uniform,
    init_recurrent_kernel = glorot_uniform,
    bias = true)

Recurrent Additive Network cell (Lee et al., 2017). See RAN for a layer that processes entire sequences.

Arguments

input_size => hidden_size: input and inner dimension of the layer.

Keyword arguments

init_kernel: initializer for the input to hidden weights. Default is glorot_uniform.

init_recurrent_kernel: initializer for the hidden to hidden weights. Default is glorot_uniform.

bias: include a bias or not. Default is true.

Equations

\[\begin{aligned} \tilde{\mathbf{c}}(t) &= \mathbf{W}^{c}_{ih} \mathbf{x}(t) + \mathbf{b}^{c}, \\ \mathbf{i}(t) &= \sigma\left( \mathbf{W}^{i}_{ih} \mathbf{x}(t) + \mathbf{W}^{i}_{hh} \mathbf{h}(t-1) + \mathbf{b}^{i} \right), \\ \mathbf{f}(t) &= \sigma\left( \mathbf{W}^{f}_{ih} \mathbf{x}(t) + \mathbf{W}^{f}_{hh} \mathbf{h}(t-1) + \mathbf{b}^{f} \right), \\ \mathbf{c}(t) &= \mathbf{i}(t) \odot \tilde{\mathbf{c}}(t) + \mathbf{f}(t) \odot \mathbf{c}(t-1), \\ \mathbf{h}(t) &= g\left( \mathbf{c}(t) \right) \end{aligned}\]

Forward

rancell(inp, (state, cstate))
rancell(inp)

Arguments

inp: The input to the rancell. It should be a vector of size input_size or a matrix of size input_size x batch_size.

(state, cstate): A tuple containing the hidden and cell states of the RANCell. They should be vectors of size hidden_size or matrices of size hidden_size x batch_size. If not provided, they are assumed to be vectors of zeros, initialized by Flux.initialstates.

Returns

A tuple (output, state), where output = new_state is the new hidden state and state = (new_state, new_cstate) is the new hidden and cell state. They are tensors of size hidden_size or hidden_size x batch_size.