TRNNCell

RecurrentLayers.TRNNCell — Type

TRNNCell(input_size => hidden_size, [activation];
    init_kernel = glorot_uniform,
    init_recurrent_kernel = glorot_uniform,
    bias = true)

Strongly typed recurrent unit (Balduzzi and Ghifary, 20–22 Jun 2016). See TRNN for a layer that processes entire sequences.

Arguments

input_size => hidden_size: input and inner dimension of the layer.

activation: activation function. Default is tanh.

Keyword arguments

init_kernel: initializer for the input to hidden weights. Default is glorot_uniform.

init_recurrent_kernel: initializer for the hidden to hidden weights. Default is glorot_uniform.

bias: include a bias or not. Default is true.

Equations

\[\begin{aligned} \mathbf{z}(t) &= \mathbf{W}_{ih} \mathbf{x}(t) + \mathbf{b}^{z} \\ \mathbf{f}(t) &= \sigma\left( \mathbf{W}_{fh} \mathbf{x}(t) + \mathbf{b}^{f} \right) \\ \mathbf{h}(t) &= \mathbf{f}(t) \odot \mathbf{h}(t-1) + \left(1 - \mathbf{f}(t)\right) \odot \mathbf{z}(t) \end{aligned}\]

Forward

trnncell(inp, state)
trnncell(inp)

Arguments

inp: The input to the trnncell. It should be a vector of size input_size or a matrix of size input_size x batch_size.

state: The hidden state of the TRNNCell. It should be a vector of size hidden_size or a matrix of size hidden_size x batch_size. If not provided, it is assumed to be a vector of zeros, initialized by Flux.initialstates.

Returns

A tuple (output, state), where both elements are given by the updated state new_state, a tensor of size hidden_size or hidden_size x batch_size.