JANET

RecurrentLayers.JANET — Type

JANET(input_size => hidden_size;
    return_state = false, kwargs...)

Just another network (van der Westhuizen and Lasenby, 2018). See JANETCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer.

Keyword arguments

init_kernel: initializer for the input to hidden weights. Default is glorot_uniform.

init_recurrent_kernel: initializer for the hidden to hidden weights. Default is glorot_uniform.

bias: include a bias or not. Default is true.

return_state: Option to return the last state together with the output. Default is false.

beta_value: control over the input data flow. Default is 1.0.

Equations

\[\begin{aligned} \mathbf{s}(t) &= \mathbf{W}^{f}_{hh} \mathbf{h}(t-1) + \mathbf{W}^{f}_{ih} \mathbf{x}(t) + \mathbf{b}^{f}, \\ \tilde{\mathbf{c}}(t) &= \tanh\left( \mathbf{W}^{c}_{hh} \mathbf{h}(t-1) + \mathbf{W}^{c}_{ih} \mathbf{x}(t) + \mathbf{b}^{c} \right), \\ \mathbf{c}(t) &= \sigma\left( \mathbf{s}(t) \right) \odot \mathbf{c}(t-1) + \left( 1 - \sigma\left( \mathbf{s}(t) - \beta \right) \right) \odot \tilde{\mathbf{c}}(t), \\ \mathbf{h}(t) &= \mathbf{c}(t). \end{aligned}\]

Forward

janet(inp, (state, cstate))
janet(inp)

Arguments

inp: The input to the janet. It should be a vector of size input_size x len or a matrix of size input_size x len x batch_size.

(state, cstate): A tuple containing the hidden and cell states of the JANET. They should be vectors of size hidden_size or matrices of size hidden_size x batch_size. If not provided, they are assumed to be vectors of zeros, initialized by Flux.initialstates.

Returns

New hidden states new_states as an array of size hidden_size x len x batch_size. When return_state = true it returns a tuple of the hidden stats new_states and the last state of the iteration.