RAN

RecurrentLayers.RAN — Type

RAN(input_size => hidden_size;
    return_state = false, kwargs...)

Recurrent Additive Network cell (Lee et al., 2017). See RANCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer.

Keyword arguments

init_kernel: initializer for the input to hidden weights. Default is glorot_uniform.

init_recurrent_kernel: initializer for the hidden to hidden weights. Default is glorot_uniform.

bias: include input to recurrent bias or not. Default is true.

recurrent_bias: include recurrent to recurrent bias or not. Default is true.

independent_recurrence: flag to toggle independent recurrence. If true, the recurrent to recurrent weights are a vector instead of a matrix. Default false.

integration_mode: determines how the input and hidden projections are combined. The options are :addition and :multiplicative_integration. Defaults to :addition.

return_state: Option to return the last state together with the output. Default is false.

Equations

\[\begin{aligned} \tilde{\mathbf{c}}(t) &= \mathbf{W}^{c}_{ih} \mathbf{x}(t) + \mathbf{b}^{c} \\ \mathbf{i}(t) &= \sigma\left( \mathbf{W}^{i}_{ih} \mathbf{x}(t) + \mathbf{W}^{i}_{hh} \mathbf{h}(t-1) + \mathbf{b}^{i} \right) \\ \mathbf{f}(t) &= \sigma\left( \mathbf{W}^{f}_{ih} \mathbf{x}(t) + \mathbf{W}^{f}_{hh} \mathbf{h}(t-1) + \mathbf{b}^{f} \right) \\ \mathbf{c}(t) &= \mathbf{i}(t) \odot \tilde{\mathbf{c}}(t) + \mathbf{f}(t) \odot \mathbf{c}(t-1) \\ \mathbf{h}(t) &= g\left( \mathbf{c}(t) \right) \end{aligned}\]

Forward

ran(inp, (state, cstate))
ran(inp)

Arguments

inp: The input to the ran. It should be a vector of size input_size x len or a matrix of size input_size x len x batch_size.

(state, cstate): A tuple containing the hidden and cell states of the RAN. They should be vectors of size hidden_size or matrices of size hidden_size x batch_size. If not provided, they are assumed to be vectors of zeros, initialized by Flux.initialstates.

Returns

New hidden states new_states as an array of size hidden_size x len x batch_size. When return_state = true it returns a tuple of the hidden stats new_states and the last state of the iteration.