TLSTM

RecurrentLayers.TLSTM — Type

TLSTM(input_size => hidden_size;
    return_state = false, kwargs...)

Strongly typed long short term memory (Balduzzi and Ghifary, 20–22 Jun 2016). See TLSTMCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer.

Keyword arguments

return_state: Option to return the last state together with the output. Default is false.

init_kernel: initializer for the input to hidden weights. Default is glorot_uniform.

init_recurrent_kernel: initializer for the hidden to hidden weights. Default is glorot_uniform.

bias: include a bias or not. Default is true.

Equations

\[\begin{aligned} \mathbf{z}(t) &= \mathbf{W}^{z}_{ih} \mathbf{x}(t) + \mathbf{W}^{z}_{hh} \mathbf{x}(t-1) + \mathbf{b}^{z} \\ \mathbf{f}(t) &= \sigma\left( \mathbf{W}^{f}_{ih} \mathbf{x}(t) + \mathbf{W}^{f}_{hh} \mathbf{x}(t-1) + \mathbf{b}^{f} \right) \\ \mathbf{o}(t) &= \tau\left( \mathbf{W}^{o}_{ih} \mathbf{x}(t) + \mathbf{W}^{o}_{hh} \mathbf{x}(t-1) + \mathbf{b}^{o} \right) \\ \mathbf{c}(t) &= \mathbf{f}(t) \odot \mathbf{c}(t-1) + \left(1 - \mathbf{f}(t)\right) \odot \mathbf{z}(t) \\ \mathbf{h}(t) &= \mathbf{c}(t) \odot \mathbf{o}(t) \end{aligned}\]

Forward

tlstm(inp, state)
tlstm(inp)

Arguments

inp: The input to the tlstm. It should be a vector of size input_size x len or a matrix of size input_size x len x batch_size.

state: The hidden state of the TLSTM. If given, it is a vector of size hidden_size or a matrix of size hidden_size x batch_size. If not provided, it is assumed to be a vector of zeros, initialized by Flux.initialstates.

Returns

New hidden states new_states as an array of size hidden_size x len x batch_size. When return_state = true it returns a tuple of the hidden states new_states and the last state of the iteration.