Cell wrappers

RecurrentLayers.RAN — Type

RAN(input_size => hidden_size; kwargs...)

The RANCell, introduced in this paper, is a recurrent cell layer which provides additional memory through the use of gates.

and returns both ht anf ct.

See RANCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.IndRNN — Type

IndRNN((input_size, hidden_size)::Pair, σ = tanh, σ=relu;
    kwargs...)

Independently recurrent network. See IndRNNCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

σ: activation function. Default is tanh

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.LightRU — Type

LightRU((input_size => hidden_size)::Pair; kwargs...)

Light recurrent unit network. See LightRUCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.LiGRU — Type

LiGRU((input_size => hidden_size)::Pair; kwargs...)

Light gated recurrent network. The implementation does not include the batch normalization as described in the original paper. See LiGRUCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.MGU — Type

MGU((input_size => hidden_size)::Pair; kwargs...)

Minimal gated unit network. See MGUCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.NAS — Type

NAS((input_size => hidden_size)::Pair; kwargs...)

Neural Architecture Search unit. See NASCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.RHN — Type

RHN((input_size => hidden_size)::Pair depth=3; kwargs...)

Recurrent highway network. See RHNCellUnit for a the unit component of this layer. See RHNCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

depth: depth of the recurrence. Default is 3

couple_carry: couples the carry gate and the transform gate. Default true

init_kernel: initializer for the input to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.MUT1 — Type

MUT1((input_size => hidden_size); kwargs...)

Mutated unit 1 network. See MUT1Cell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.MUT2 — Type

MUT2Cell((input_size => hidden_size); kwargs...)

Mutated unit 2 network. See MUT2Cell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.MUT3 — Type

MUT3((input_size => hidden_size); kwargs...)

Mutated unit 3 network. See MUT3Cell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

RecurrentLayers.SCRN — Type

SCRN((input_size => hidden_size)::Pair;
    init_kernel = glorot_uniform,
    init_recurrent_kernel = glorot_uniform,
    bias = true,
    alpha = 0.0)

Structurally contraint recurrent unit. See SCRNCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

alpha: structural contraint. Default is 0.0

RecurrentLayers.PeepholeLSTM — Type

PeepholeLSTM((input_size => hidden_size)::Pair; kwargs...)

Peephole long short term memory network. See PeepholeLSTMCell for a layer that processes a single sequence.

Arguments

input_size => hidden_size: input and inner dimension of the layer

init_kernel: initializer for the input to hidden weights

init_recurrent_kernel: initializer for the hidden to hidden weights

bias: include a bias or not. Default is true

Equations

\[\begin{align} f_t &= \sigma_g(W_f x_t + U_f c_{t-1} + b_f), \\ i_t &= \sigma_g(W_i x_t + U_i c_{t-1} + b_i), \\ o_t &= \sigma_g(W_o x_t + U_o c_{t-1} + b_o), \\ c_t &= f_t \odot c_{t-1} + i_t \odot \sigma_c(W_c x_t + b_c), \\ h_t &= o_t \odot \sigma_h(c_t). \end{align}\]